Split out `TempTableName` tests
