+++ clients/mssql/store_test.go
@@ -1,8 +1,12 @@
 package mssql
 
 import (
+	"strconv"
+	"strings"
 	"testing"
+	"time"
 
+	"github.com/artie-labs/transfer/clients/shared"
 	"github.com/artie-labs/transfer/lib/config"
 	"github.com/artie-labs/transfer/lib/kafkalib"
 	"github.com/artie-labs/transfer/lib/optimization"
@@ -36,4 +40,19 @@ func TestFullyQualifiedName(t *testing.T) {
 		assert.Equal(t, `schema."table"`, store.ToFullyQualifiedName(tableData, true), "escaped")
 		assert.Equal(t, "schema.table", store.ToFullyQualifiedName(tableData, false), "unescaped")
 	}
+}
+
+func TestTempTableName(t *testing.T) {
+	trimTTL := func(tableName string) string {
+		lastUnderscore := strings.LastIndex(tableName, "_")
+		assert.GreaterOrEqual(t, lastUnderscore, 0)
+		epoch, err := strconv.ParseInt(tableName[lastUnderscore+1:], 10, 64)
+		assert.NoError(t, err)
+		assert.Greater(t, time.Unix(epoch, 0), time.Now().Add(5*time.Hour)) // default TTL is 6 hours from now
+		return tableName[:lastUnderscore]
+	}
+
+	tableData := optimization.NewTableData(nil, config.Replication, nil, kafkalib.TopicConfig{Database: "db", Schema: "schema"}, "table")
+	tempTableName := shared.TempTableName(&Store{}, tableData, "sUfFiX")
+	assert.Equal(t, "schema.table___artie_sUfFiX", trimTTL(tempTableName))
 }
